requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
mplfinance>=0.12.10b0
matplotlib>=3.8.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
